new commit changes

+++ app/Http/Controllers/ListingController.php
@@ -5,36 +5,31 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Storage;
 
 class ListingController extends Controller
 {
-    //all listings
-    public function index()
-    {
+    // Show all listings
+    public function index() {
         return view('listings.index', [
-        'heading' => 'Latest Listing',
-        // 'listings' => Listing::all(),
-    'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
-    // 'listings' => Listing::latest()->filter(request(['tag', 'search']))->simplePaginate(6)
- ]);
+            'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
+        ]);
     }
 
-    //single listings
-    public function show(Listing $listing)
-    {
+    //Show single listing
+    public function show(Listing $listing) {
         return view('listings.show', [
-      'listing' => $listing
-    ]);
+            'listing' => $listing
+        ]);
     }
 
     // Show Create Form
-    public function create()
-    {
+    public function create() {
         return view('listings.create');
     }
+
     // Store Listing Data
-    public function store(Request $request)
-    {
+    public function store(Request $request) {
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required', Rule::unique('listings', 'company')],
@@ -45,7 +40,7 @@ class ListingController extends Controller
             'description' => 'required'
         ]);
 
-        if ($request->hasFile('logo')) {
+        if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
@@ -56,16 +51,18 @@ class ListingController extends Controller
         return redirect('/')->with('message', 'Listing created successfully!');
     }
 
-
     // Show Edit Form
-    public function edit(Listing $listing)
-    {
+    public function edit(Listing $listing) {
         return view('listings.edit', ['listing' => $listing]);
     }
 
     // Update Listing Data
-    public function update(Request $request, Listing $listing)
-    {
+    public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -76,7 +73,7 @@ class ListingController extends Controller
             'description' => 'required'
         ]);
 
-        if ($request->hasFile('logo')) {
+        if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
@@ -84,14 +81,23 @@ class ListingController extends Controller
 
         return back()->with('message', 'Listing updated successfully!');
     }
-    // Delete Listing
-    public function destroy(Listing $listing)
-    {
 
-        // if($listing->logo && Storage::disk('public')->exists($listing->logo)) {
-        //     Storage::disk('public')->delete($listing->logo);
-        // }
+    // Delete Listing
+    public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
+        if($listing->logo && Storage::disk('public')->exists($listing->logo)) {
+            Storage::disk('public')->delete($listing->logo);
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
